requirements.txt
accelerate==1.10.1
certifi==2025.10.5
diffusers==0.35.1
hf-xet==1.1.10
huggingface-hub==0.35.3
numcodecs==0.12.1
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
psutil==7.1.0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
torch==2.8.0
tqdm==4.67.1
triton==3.4.0
urllib3==2.5.0
zarr==2.18.2
zipp==3.23.0
